Data1D's y is an unsigned int

+++ src/data/Data1DIterator.cpp
@@ -4,6 +4,6 @@ inline float Data1DIterator::getX() const {
     return static_cast<Data1D*>(this->iterator.get())->getValue();
 }
 
-inline float Data1DIterator::getY() const {
+inline unsigned Data1DIterator::getY() const {
     return static_cast<Data1D*>(this->iterator.get())->getCount();
 }
